Switch getDualSignature to avoiding interpreter database
Now that OpenAxiom emits the dual signature data also directly in the
generated executable code for a constructor, there is no no need to
consult the interpreter data base for that information.  This change
relies on the previous commit, e.g. getDualSignature should be
considered invalid for builtin constructors.  Any instanbility caused
by this patch should be considered a latent bug that should be given
high priority fixing.