getConstructorAncestorsFromDB: get them from generated code
Ancestors for category constructors are stored in databases with
parameter names of the form #1, #2, etc. called "formals", instead of
the user-writtern parameters in the input souce code.  However, the
ancestors as previously stored in generated code used the input
source-level parameter names.  Changed to formals, so no rewrite
needed before making databases.