Better indirect call resolution
The infrastructure put in place in the last few commits now enables
better indirect calls ('SPADCALL') to operations implemented in the
same capsule.  The improvement here is that the limitation of
unconditional definition is removed.  Now, even conditional
definitions are resolved as long as the its predicate and the
predicate of the current function match.