Sequester Hyper `StateNode`
Move `StateNode` to `hyper/lex.c` as the only place where it is referenced.

Rename `StateNode` to `IOState`.  It is not a node.

Introduce `IOStateManager` as a RAII type for reliable save/restore mechanism.  Remove `save_scanner_state()` and `restore_scanner_state()`

Introduce `IOStateStack` as a non-intrusive stack datatype based on C++ standard `std::stack`.

Also save `line_number` which was forgotten as part of the IO state to save.